Final Commit
This commit shows that all of my files were modified. This is not accurate. For this last change (since I did some other stuff with some of my files), I used my last push as
a restore point and made some modifications from there. I downloaded the ZIP from GitHub and then modified 3 files. I modified utilityFunctions.c to include a new toLower() function
that lowers the case of a string that is passed to it, and I modified it to replace all instances of "strlwr" with "toLower". I modified utilityFunctions.h to include a function
prototype for the previously mentioned toLower() function. I also modified crudFunctions.c to replace the one instance of "strlwr" with "toLower". With these changes, this program
now compiles and runs in both Windows and Ubuntu. These changes (and the absence of changes in files I did not mention were changed) can be confirmed by comparing the files in this
commit to those in the last commit.